backend/requirements.txt
alembic==1.16.4
amqp==5.3.1
aniso8601==10.0.1
attrs==25.3.0
babel==2.17.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.1
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
greenlet==3.2.3
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
omise==0.11.0
python-dotenv==1.0.0
flask-cors==4.0.0
